agents/gemini_agent.py
import os
from typing import Dict, Any, List, Union
from core.base_agent import BaseAgent
import google.generativeai as genai
import numpy as np
from PIL import Image
import io

class GeminiAgent(BaseAgent):
    def __init__(self):
        super().__init__("Gemini", "Multi-Modal Expert")
        genai.configure(api_key=os.getenv("GOOGLE_API_KEY"))
        self.model = genai.GenerativeModel('gemini-pro-vision')
        self.text_model = genai.GenerativeModel('gemini-pro')
        self.vision_embeddings = {}
        self.text_embeddings = {}
        self.cross_modal_adapter = None
        
    async def initialize(self) -> None:
        """Initialize multi-modal components"""
        # Initialize vision embeddings
        self.vision_embeddings = self._initialize_vision_embeddings()
        
        # Initialize text embeddings
        self.text_embeddings = self._initialize_text_embeddings()
        
        # Initialize cross-modal adapter
        self.cross_modal_adapter = self._create_cross_modal_adapter()
        
        self.log_info("Initialized multi-modal components")
        
    async def process_task(self, task: Dict[str, Any]) -> Dict[str, Any]:
        """Process multi-modal task"""
        try:
            # Determine task modalities
            modalities = self._identify_modalities(task)
            
            # Process each modality
            modal_results = {}
            for modality in modalities:
                if modality == "vision":
                    modal_results[modality] = await self._process_vision(task)
                elif modality == "text":
                    modal_results[modality] = await self._process_text(task)
                    
            # Cross-modal fusion
            fused_result = self._fuse_modalities(modal_results)
            
            return {
                "modalities": modalities,
                "modal_results": modal_results,
                "fused_result": fused_result
            }
            
        except Exception as e:
            self.log_error(f"Error in multi-modal processing: {str(e)}")
            return {"error": str(e)}
            
    async def collaborate(self, agent_outputs: Dict[str, Any]) -> Dict[str, Any]:
        """Refine multi-modal understanding based on other agents"""
        try:
            # Analyze modal contributions
            modal_analysis = self._analyze_modal_contributions(agent_outputs)
            
            # Adjust cross-modal weights
            adjusted_weights = self._adjust_cross_modal_weights(modal_analysis)
            
            # Update embeddings
            self._update_embeddings(modal_analysis)
            
            return {
                "modal_analysis": modal_analysis,
                "weight_adjustments": adjusted_weights,
                "embedding_updates": self._get_embedding_status()
            }
            
        except Exception as e:
            self.log_error(f"Error in multi-modal collaboration: {str(e)}")
            return {"error": str(e)}
            
    async def adapt(self, feedback: Dict[str, Any]) -> None:
        """Adapt multi-modal processing based on feedback"""
        try:
            if "performance_metrics" in feedback:
                # Update vision processing
                self._update_vision_processing(feedback["performance_metrics"])
                
                # Update text processing
                self._update_text_processing(feedback["performance_metrics"])
                
                # Update cross-modal fusion
                self._update_fusion_strategy(feedback["performance_metrics"])
                
                self.log_info("Multi-modal components adapted based on feedback")
                
        except Exception as e:
            self.log_error(f"Error in multi-modal adaptation: {str(e)}")
            
    def _initialize_vision_embeddings(self) -> Dict[str, np.ndarray]:
        """Initialize vision embeddings"""
        return {
            "object": np.random.randn(1024, 256),
            "scene": np.random.randn(1024, 256),
            "action": np.random.randn(1024, 256)
        }
        
    def _initialize_text_embeddings(self) -> Dict[str, np.ndarray]:
        """Initialize text embeddings"""
        return {
            "semantic": np.random.randn(768, 256),
            "syntactic": np.random.randn(768, 256),
            "contextual": np.random.randn(768, 256)
        }
        
    def _create_cross_modal_adapter(self) -> Any:
        """Create cross-modal adaptation mechanism"""
        class CrossModalAdapter:
            def __init__(self):
                self.vision_projection = np.random.randn(1024, 512)
                self.text_projection = np.random.randn(768, 512)
                self.fusion_layer = np.random.randn(512, 256)
                
            def fuse(self, vision_feat: np.ndarray, text_feat: np.ndarray) -> np.ndarray:
                vision_proj = np.dot(vision_feat, self.vision_projection)
                text_proj = np.dot(text_feat, self.text_projection)
                fused = np.tanh(vision_proj + text_proj)
                return np.dot(fused, self.fusion_layer)
                
        return CrossModalAdapter()
        
    def _identify_modalities(self, task: Dict[str, Any]) -> List[str]:
        """Identify task modalities"""
        modalities = []
        if "image" in task or "vision" in task:
            modalities.append("vision")
        if "text" in task or "query" in task:
            modalities.append("text")
        return modalities
        
    async def _process_vision(self, task: Dict[str, Any]) -> Dict[str, Any]:
        """Process vision input"""
        if "image" not in task:
            return {"error": "No image provided"}
            
        try:
            # Process image with Gemini
            response = await self.model.generate_content(
                ["Analyze this image:", task["image"]]
            )
            
            # Extract vision features
            vision_features = self._extract_vision_features(task["image"])
            
            return {
                "gemini_analysis": response.text,
                "vision_features": vision_features.tolist()
            }
            
        except Exception as e:
            return {"error": f"Vision processing error: {str(e)}"}
            
    async def _process_text(self, task: Dict[str, Any]) -> Dict[str, Any]:
        """Process text input"""
        try:
            # Process text with Gemini
            response = await self.text_model.generate_content(
                task.get("text", task.get("query", ""))
            )
            
            # Extract text features
            text_features = self._extract_text_features(task)
            
            return {
                "gemini_analysis": response.text,
                "text_features": text_features.tolist()
            }
            
        except Exception as e:
            return {"error": f"Text processing error: {str(e)}"}
            
    def _fuse_modalities(self, modal_results: Dict[str, Any]) -> Dict[str, Any]:
        """Fuse results from different modalities"""
        try:
            vision_feat = np.array(modal_results.get("vision", {}).get("vision_features", []))
            text_feat = np.array(modal_results.get("text", {}).get("text_features", []))
            
            if len(vision_feat) > 0 and len(text_feat) > 0:
                fused = self.cross_modal_adapter.fuse(vision_feat, text_feat)
                return {"fused_features": fused.tolist()}
            else:
                return {"error": "Insufficient features for fusion"}
                
        except Exception as e:
            return {"error": f"Fusion error: {str(e)}"}
            
    def _extract_vision_features(self, image: Union[str, Image.Image]) -> np.ndarray:
        """Extract features from vision input"""
        # Implement vision feature extraction
        return np.random.randn(1024)  # Placeholder
        
    def _extract_text_features(self, task: Dict[str, Any]) -> np.ndarray:
        """Extract features from text input"""
        # Implement text feature extraction
        return np.random.randn(768)  # Placeholder
