requirements.txt
torch>=2.1.0
transformers>=4.36.0
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0
tqdm>=4.65.0
matplotlib>=3.8.0
openai>=1.59.8
anthropic>=0.43.1
google-generativeai>=0.8.3
cohere>=5.13.8
groq>=0.15.0
mistralai==0.4.2
crewai>=0.12.1
httpx>=0.23.0
pydantic>=1.9.0
python-dotenv>=0.19.0
requests>=2.31.0
typing-extensions>=4.11.0
aiohttp>=3.8.0
asyncio>=3.4.3
beautifulsoup4>=4.12.0
selenium>=4.15.0
webdriver-manager>=4.0.1
fastapi>=0.105.0
uvicorn>=0.24.0
pytest>=7.4.0
